si-gun-gu API 수정

+++ src/main/java/com/carnival/matchcube/dto/AreaDTO.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 @Getter
 @Setter
 public class AreaDTO {
-    public Long code;
+    public short code;
 }
